+++ package.json
@@ -9,9 +9,7 @@
   },
   "dependencies": {
     "firebase": "^9.18.0",
-
-    "firebaseui": "^6.0.2",
-
+    "firebaseui": "^0.600.0",
     "vue": "^3.2.47",
     "vue-router": "^4.1.6"
   },
