+++ src/firebase.js
@@ -9,9 +9,11 @@ const firebaseConfig = {
   projectId: "bt3103-74add",
   storageBucket: "bt3103-74add.appspot.com",
   messagingSenderId: "712226669576",
-  appId: "1:712226669576:web:750d8f4eccdef2791e7dbd"
+  appId: "1:712226669576:web:750d8f4eccdef2791e7dbd",
+  measurementId: "G-Z9QDX69313"
 };
 
-// Initialize Firebase
+// Initialize Firebase, databse and auth
 const firebaseApp = initializeApp(firebaseConfig);
+
 export default firebaseApp;
