+++ src/views/ForgotPassword.vue
@@ -1,56 +1,81 @@
-<script setup>
-import Popup from "../components/Popup.vue";
-</script>  
 
 <template>
-    <div class ='reset-password'>
-        <Popup v-if= 'popupActive' @close='close'/>
+    <div class="reset-password">
+        <Modal v-if="modalActive" :modalMessage="modalMessage" @close-modal="closeModal" />
         <div class="form-wrap">
-        <form class='reset' action="">
-            <h2>Reset Password</h2>
-            <p>Forgot password? Please enter your email</p>   
-            <div class="inputs">
-                <div class="input">
-                    <input type="text" placeholder = "Email" v-model="email"/>
+            <form class="reset">
+                <p class="login-register">
+                    Already have an account?
+                    <router-link class="router-link" :to="{ name: 'Login' }">Login</router-link>
+                </p>
+                <h2>Reset Password</h2>
+                <p>Forgot your password? Enter your email to reset it</p>
+                <div class="inputs">
+                    <div class="input">
+                        <input type="text" placeholder="Email" v-model="email" />
+                    </div>
                 </div>
-            </div>
-
-            <button>Reset Password</button> 
-        </form>
+                <button @click.prevent="resetPassword"> Reset </button>
+            </form>
         </div>
     </div>
 </template>
-
+  
 <script>
+import Modal from "../components/Modal.vue";
+import { sendPasswordResetEmail, getAuth } from "firebase/auth";
+import firebaseApp from '@/firebase.js';
+
+const auth = getAuth(firebaseApp);
 
 export default {
-    name:'ForgotPassword',
-    data(){
+    name: "ForgotPassword",
+    data() {
         return {
-            email:null,
-            popupActive:null,
-            popupMessage:"",
-        }
+            email: null,
+            modalActive: false,
+            modalMessage: "",
+        };
     },
 
-    methods:{
-        close(){
-            this.popupActive = !this.popupActive;
-            this.email = '';
-        }
-    }
-}
-</script>
+    components: {
+        Modal,
+    },
 
-<style scoped>
+    methods: {
+        resetPassword() {
+            sendPasswordResetEmail(auth, this.email).then(() => {
+                this.modalMessage = "If your account exists, you will receive an email";
+                this.modalActive = true;
+            }).catch((err) => {
+                this.modalMessage = err.message;
+                this.modalActive = true
+            })
+        },
+        closeModal() {
+            this.modalActive = !this.modalActive;
+            this.email = "";
+        },
+    },
+};
+</script>
+  
+<style lang="scss" scoped>
 .reset-password {
-    position:relative;
+    position: relative;
+
     .form-wrap {
         .reset {
-            h2{
-                margin-bottom:8px;
+            h2 {
+                margin-bottom: 8px;
+            }
+
+            p {
+                text-align: center;
+                margin-bottom: 32px;
             }
         }
     }
 }
-</style>   
+</style>
+
